Entirely inspect how the stack use case work layer by layer
Introducing a complete description on how things works
and their places (position) in the use case and who call who in order
from the nova puppet project to the oslo.messaging rabbitmq driver
initialization and the heartbeat launch. (still in development/writing).

We can use/reuse this analyze to write the blog post in a second time.